dasher: add todo list and clean up nebula code

+++ raylib/projects/dapper-dasher/src/includes/player.h
@@ -5,13 +5,13 @@
 
 class Player{
     private:
-        
+
     public:
         const int WIDTH{128}; // Player width
         const int HEIGHT{128}; // Player height
         // const int BASE_SPEED{2}; // Player normal speed
         const int BASE_SPEED{200}; // Player normal speed (pixes/second)
-        const int JUMP_HEIGHT{this->HEIGHT*1.3}; // Player jump height
+        const int JUMP_HEIGHT{int(this->HEIGHT*1.3)}; // Player jump height
         const float UPDATE_TIME{1.0/12.0}; // Ammount of time that should pass between each animation frame. 1/12 will update the frame 12 times per second.
 
         // Texture sprite sheet. Must be loaded after the Window is created.
